Split up checks into syntax and expression folders
