+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-arg-count.ts
@@ -2,10 +2,9 @@ import { msgid, ngettext } from "ttag";
 
 import * as Lib from "metabase-lib";
 
-import { getClauseDefinition } from "../config";
-import { visit } from "../visitor";
-
-import { error } from "./utils";
+import { getClauseDefinition } from "../../config";
+import { visit } from "../../visitor";
+import { error } from "../utils";
 
 export function checkArgCount({
   expressionParts,
