+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-case-or-if-arg-count.ts
@@ -2,9 +2,8 @@ import { t } from "ttag";
 
 import * as Lib from "metabase-lib";
 
-import { visit } from "../visitor";
-
-import { error } from "./utils";
+import { visit } from "../../visitor";
+import { error } from "../utils";
 
 export function checkCaseOrIfArgCount({
   expressionParts,
