+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-lib-diagnostics.ts
@@ -1,6 +1,6 @@
 import * as Lib from "metabase-lib";
 
-import { DiagnosticError } from "../errors";
+import { DiagnosticError } from "../../errors";
 
 export function checkLibDiagnostics({
   query,
