+++ frontend/src/metabase-lib/v1/expressions/diagnostics/syntax/check-missing-comma-in-argument-list.ts
@@ -1,6 +1,6 @@
 import { t } from "ttag";
 
-import type { Token } from "../pratt";
+import type { Token } from "../../pratt";
 import {
   CALL,
   FIELD,
@@ -8,9 +8,8 @@ import {
   GROUP_CLOSE,
   IDENTIFIER,
   OPERATORS,
-} from "../pratt";
-
-import { error } from "./utils";
+} from "../../pratt";
+import { error } from "../utils";
 
 export function checkMissingCommasInArgumentList({
   tokens,
